requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tornado==6.4.1